:truck: Rename: 폴더명 camel -> kebab case 적용
